Refactor parser methods for commands to CommandParser classes
