+++ src/test/java/parser/ParserTest.java
@@ -53,6 +53,7 @@ class ParserTest {
     @Test
     public void parse_backCommandWithArgs_expectException() {
         Access access = new Access();
+        access.setIsModuleLevel();
         final String[] inputs = {"back args"};
         for (String input : inputs) {
             assertThrows(InvalidInputException.class, () -> Parser.parse(input, access));
